OR-534 Resolve the unread problem when activity stream is empty
Before commit : when an activity stream is empty and when the user never
do GET /api/activitystream/:uuid, the tracker return always 0 unread
message.

After commit : when an activity stream is empty and when the user never
do GET /api/activitystream/:uuid, the tracker query the activity stream
in order to count number of timeline entries.