OR-534 When a new user join a community, unread message is 0
When a user join a community, query the activity stream in order to
create the timeline entries tracker. If there is no timeline entry the
tracker is not create and if there are one or more timeline entries then
the tracker is create and set to the last timeline entry.