OR-751 remove inSlicesOf filter
This is a really sad move, because the filter approach was way more elegant.
However, because Angular uses strict equality comparison, and we return arrays,
we are not able to bring a working filter without a global cache.

And the global cache brings infinite digest cycles error when we instanciate
several directives using the inSlicesOf filter.