OR-1097 Select main tab item by sniffing URL hash
This commit removes the old manual tab item selection, and uses the URL hash to
choose what item to select. If you go to the URL : #/something/great/inside ,
the directive will look for an element that has class .esn-item and a property
data-esn-path='something' . If found the class "active" is added to the element.