ADM-86 Fix esn-config backward compatibility
The idea is to find documents in 'configuration', 'features' and
'configurations' then merge to one (merged document). By reading from this
merged document, we can make sure that fallback compatibility always works.
I will write unit-storage tests to prove that it works.

While 'configuration' and 'features' collections are read-only, that means we
can cache it (currently in memory, later in cache system when we have it) to
improve performance.