CAL-50 recurring form: force interval to 1
Before this PR it was possible to create an recurring event with in
interval of 0.
Now when the user made an event a recurring event the input interval is
prefilled with 1 and can't be modified to enter something lower than
that.
Also when the user change the freq of an event the interval value is
preserved if valid, this was not the case before.