CAL-269 calendarshell: apply reply jCal
When issuing a iTipMessage Sabre (acting according the RFC) sent a iTip
that contain a vevent which does not comport all data about original
event (for example it can miss RRULE and it miss other participant
partstat). This method aim to update original CalendarShell by taking
revelent info of the jCal contained in the iTipAnswer.