linagora.esn.contact#350 workaround to fix issue cannot update contact 2 times sequentially
When the user clicks Save to update contact, the contact is not really
saved to database because of graceperiod timeout (usually 8s). To
display updated contact to user, we use `contactUpdateDataService` to
cache contact after it is saved from contact edition page (see controller.js line 284)

Although the information of cached contact is up to date, the ETag of it
is outdated, so when the user tries to save contact again, it is failed
because ETag is not matched.

To workaround this, we listen to CONTACT_EVENTS.UPDATED to update the
ETag, we use it in 2 places:

- In `editContactController`: for the case user edits contact before
graceperiod ends
- In `contactUpdateDataService`: (in this PR) for the case user edit after
graceperiod ends