#3049 OpenId Connect: generic implementation
Previous implementation was taking for granted that the access_token is
a JWT providing meta-information on the provider. This is not the case
for LemonLDAP, which sends back a random string.

This implementation will then test the access_token against all defined
OIDC providers, and returns the information of the first that accepts
the toekn.